Split registry_ module to many modules
