+++ lib/registry_/Node.js
@@ -1,118 +1,13 @@
 var inherit = require('inherit'),
-    Set = require('collections/set'),
-    EventEmitter = require('events').EventEmitter,
     q = require('q'),
     http = require('q-io/http'),
-    async = require('./q-async'),
-    extend = require('extend'),
+    async = require('../q-async'),
     util = require('util'),
-    log = require('./log'),
-    driver = require('./driver'),
-    matchesCaps = require('./matcher'),
-    normalizeCaps = require('./normalizeCaps');
-
-module.exports = inherit({
-    __constructor: function(config) {
-        this.config = config;
-        this.nodes = new NodeSet();
-        this.sessions = new SessionSet();
-    },
-
-    getNewSession: function(req) {
-        var self = this;
-        return self.nodes.getNewSession(req)
-            .then(function(session) {
-                if (session) {
-                    self.sessions.add(session);
-                }
-                return session;
-            });
-    },
-
-    getSessionById: function(sessionId) {
-        return this.sessions.getById(sessionId);
-    },
-
-    createNode: function(json) {
-        return new Node(this, json);
-    },
-
-    add: function(node) {
-        var self = this;
-        return self.removeIfPresent(node)
-            .then(function() {
-                return self.nodes.add(node);
-            });
-    },
-
-    removeIfPresent: function(node) {
-        var self = this;
-        return self.nodes.has(node)
-            .then(function(exists) {
-                if (!exists) {
-                    return;
-                }
-                return self.nodes.getById(node.getId())
-                    .then(function(node) {
-                        return node.unregister();
-                    });
-            });
-    },
-
-    remove: function(node) {
-        return this.nodes.remove(node);
-    },
-
-    getById: function(id) {
-        return this.nodes.getById(id);
-    }
-});
-
-function _setEquals(a, b) {
-    if (typeof a !== 'string') {
-        a = a.getId();
-    }
-    if (typeof b !== 'string') {
-        b = b.getId();
-    }
-    return a === b;
-}
-
-function _setCompare(val) {
-    if (typeof val !== 'string') {
-        val = val.getId();
-    }
-    return val;
-}
-
-var NodeSet = inherit({
-    __constructor: function(config) {
-        this.config = config;
-        this.nodes = new Set(null, _setEquals, _setCompare);
-    },
-
-    getNewSession: function(req) {
-        return async.doFirstSeries(this.nodes.toArray(), function(node) {
-            return node.getNewSession(req);
-        });
-    },
-
-    getById: function(id) {
-        return q(this.nodes.get(id));
-    },
-
-    add: function(node) {
-        return q(this.nodes.add(node));
-    },
-
-    remove: function(node) {
-        return q(this.nodes.remove(node));
-    },
-
-    has: function(node) {
-        return q(this.nodes.has(node));
-    }
-});
+    extend = require('extend'),
+    EventEmitter = require('events').EventEmitter,
+    log = require('../log'),
+    normalizeCaps = require('../normalizeCaps'),
+    NodeSlot = require('./NodeSlot');
 
 const DEFAULT_POLLING_INTERVAL = 10000;
 const DEFAULT_UNREGISTER_DELAY = 60000;
@@ -131,6 +26,7 @@ var Node = inherit(EventEmitter, {
         var config = extend(true, {}, registry.config, json.configuration || {});
         this.registry = registry;
         this.config = config;
+        /** @type {Array} */
         this.caps = extend(true, json.capabilities.map(function(caps) {
             return normalizeCaps(caps)
         }));
@@ -156,6 +52,7 @@ var Node = inherit(EventEmitter, {
 
         // populate testing slots by node capabilities
         var self = this;
+        /** @type {Array} */
         this.slots = [];
         this.caps.forEach(function(caps) {
             var maxInstances = caps.maxInstances || 1;
@@ -337,134 +234,4 @@ var Node = inherit(EventEmitter, {
     }
 });
 
-var NodeSlot = inherit({
-    __constructor: function(node, caps) {
-        this.locked = false;
-        this.node = node;
-        this.caps = caps;
-        // TODO: use constants from driver
-        this.protocol = (caps.seleniumProtocol === 'Selenium') ? 'RC' : 'WebDriver';
-        this.session = null;
-    },
-
-    getNewSession: function(req) {
-        var self = this;
-        if (self.isLocked()) {
-            log.debug('Slot on node with ID=%s is locked, skip', self.node.getId());
-            return null;
-        }
-
-        var driver = self.getDriver(),
-            caps = driver._getCapabilities(req);
-
-        // match capabilities
-        if (!matchesCaps(caps, self.caps)) {
-            log.debug('Capabilities for slot on node with ID=%s do not match', self.node.getId());
-            log.debug('* desired capabilities: %j', caps);
-            log.debug('* slot capabilities: %j', self.caps);
-            return null;
-        }
-
-        this.lock();
-
-        // send open session request
-        return driver._proxyRequestToNode(req, self.node)
-            .then(function(res) {
-                // process response
-                return driver._parseProxyResponse(res)
-                    .spread(function(sessionId, resCaps) {
-                        // add session to the store
-                        //return self._addSession(self._createSession(sessionId, node, resCaps, desiredCaps));
-
-                        // TODO: save session on success
-                        var session = new Session(sessionId, self.node, resCaps, caps);
-                        self.session = session;
-                        return session;
-                    });
-            })
-            .catch(function(err) {
-                log.debug('Error ocured during session opening on node with ID=%s', self.node.getId());
-                log.debug(err.stack || err);
-
-                self.unlock();
-                return q.reject(err);
-            });
-    },
-
-    terminateSession: function() {
-        // TODO: send termination request
-        this.release();
-    },
-
-    release: function() {
-        // TODO: clear session
-        this.session = null;
-        this.unlock();
-    },
-
-    getDriver: function() {
-        return driver.getDriver(this.protocol);
-    },
-
-    lock: function() {
-        if (this.locked) {
-            throw new Error('Could not lock already locked slot.');
-        }
-        this.locked = true;
-    },
-
-    unlock: function() {
-        if (!this.locked) {
-            throw new Error('Could not unlock already unlocked slot.');
-        }
-        this.locked = false;
-    },
-
-    isLocked: function() {
-        return this.locked;
-    }
-});
-
-var Session = inherit({
-    __constructor: function(id, node, caps, desiredCaps) {
-        this.id = id;
-        this.node = node;
-        this.capabilities = caps;
-        this.desiredCapabilities = desiredCaps;
-    },
-
-    getId: function() {
-        return this.id;
-    },
-
-    toJSON: function() {
-        return {
-            id: this.getId(),
-            node: this.node.getId(),
-            capabilities: extend(true, this.capabilities),
-            desiredCapabilities: extend(true, this.desiredCapabilities)
-        }
-    }
-});
-
-var SessionSet = inherit({
-    __constructor: function() {
-        this.sessions = new Set(null, _setEquals, _setCompare);
-    },
-
-    getById: function(id) {
-        return q(this.sessions.get(id));
-    },
-
-    add: function(session) {
-        return q(this.sessions.add(session));
-    },
-
-    remove: function(session) {
-        return q(this.sessions.remove(session));
-    },
-
-    has: function(session) {
-        return q(this.sessions.has(session));
-    }
-});
+module.exports = Node;
